requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
pydantic-settings>=2.6.0
anthropic>=0.42.0
PyMuPDF>=1.25.0
pydub>=0.25.1
mlx-audio>=0.2.3
